requirements.txt
# CaseLawGPT - Requirements

# Core AI/ML
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
google-generativeai>=0.3.0
google-auth>=2.16.0

# Vector Store
pinecone-client>=3.0.0
langchain-pinecone>=0.0.1

# Text Processing
rank-bm25>=0.2.2

# API Backend - FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0

# Testing & Evaluation
pytest>=8.0.0
requests>=2.31.0
rapidfuzz>=3.9.0
